DynamicModule/requirements.txt
einops
numpy<2
Pillow
tqdm
PyYAML
setuptools
lightning==2.4.0
deepspeed
open3d
omegaconf
opencv_python
pandas
moviepy<2
av<14
timm
kornia
fairscale
diffusers
tensorboard
accelerate
optimum
autoawq